+++ skrub/_data_ops/_data_ops.py
@@ -45,6 +45,7 @@ from sklearn.base import BaseEstimator
 from .. import _config
 from .. import _dataframe as sbd
 from .. import selectors as s
+from .._apply_to_cols import ApplyToCols
 from .._check_input import cast_column_names_to_strings
 from .._reporting._utils import strip_xml_declaration
 from .._utils import PassThrough, set_module, short_repr
@@ -820,10 +821,12 @@ for op_name in _UNARY_OPS:
     setattr(DataOp, op_name, _make_unary_op(op_name))
 
 
-def _check_wrap_params(cols, how, allow_reject, reason):
+def _check_wrap_params(cols, exclude_cols, how, allow_reject, reason):
     msg = None
     if not isinstance(cols, type(s.all())):
         msg = f"`cols` must be `all()` (the default) when {reason}"
+    if exclude_cols is not None:
+        msg = f"`exclude_cols` must be None (the default) when {reason}"
     elif how not in ["auto", "no_wrap"]:
         msg = f"`how` must be 'auto' (the default) or 'no_wrap' when {reason}"
     elif allow_reject:
@@ -856,33 +859,25 @@ def _check_estimator_type(estimator):
     )
 
 
-def _check_apply_how(how):
+def _wrap_estimator(estimator, cols, exclude_cols, no_wrap, how, allow_reject, X):
+    """
+    Wrap the estimator passed to .skb.apply in ApplyToCols if needed.
+    """
+    if not isinstance(no_wrap, bool):
+        raise TypeError(
+            "The parameter 'no_wrap' of .skb.apply() must be a Boolean, "
+            f"got: {no_wrap!r}."
+        )
     valid = ["auto", "cols", "frame", "no_wrap"]
-    if how in valid:
-        return how
-
-    # TODO remove when the old names are completely dropped in 0.7.0
-    translate = {"columnwise": "cols", "sub_frame": "frame", "full_frame": "no_wrap"}
-    if how in translate:
-        new = translate[how]
+    if how not in valid:
+        raise ValueError(f"`how` must be one of {valid}. Got: {how!r}")
+    if how != "auto":
         warnings.warn(
-            (
-                f"{how!r} has been renamed to {new!r}: use .skb.apply(how={new!r})"
-                " instead."
-            ),
+            "The 'how' parameter of .skb.apply() has been deprecated "
+            "and will  be removed in a future release. "
+            f"Use the 'no_wrap' parameter instead. Got how={how!r}",
             FutureWarning,
         )
-        return new
-
-    raise ValueError(f"`how` must be one of {valid}. Got: {how!r}")
-
-
-def _wrap_estimator(estimator, cols, how, allow_reject, X):
-    """
-    Wrap the estimator passed to .skb.apply in ApplyToEachCol or ApplyToSubFrame if
-    needed.
-    """
-    how = _check_apply_how(how)
 
     if estimator in [None, "passthrough"]:
         estimator = PassThrough()
@@ -890,8 +885,11 @@ def _wrap_estimator(estimator, cols, how, allow_reject, X):
     _check_estimator_type(estimator)
 
     def _check(reason):
-        _check_wrap_params(cols, how, allow_reject, reason)
+        _check_wrap_params(cols, exclude_cols, how, allow_reject, reason)
 
+    if no_wrap:
+        _check("`no_wrap` is True")
+        return estimator
     if how == "no_wrap":
         _check("`how` is 'no_wrap'")
         return estimator
@@ -901,9 +899,17 @@ def _wrap_estimator(estimator, cols, how, allow_reject, X):
     if not sbd.is_dataframe(X):
         _check("the input is not a DataFrame")
         return estimator
-    columnwise = {"auto": "auto", "cols": True, "frame": False}[how]
+    if how == "auto":
+        return ApplyToCols(
+            estimator, cols=cols, exclude_cols=exclude_cols, allow_reject=allow_reject
+        )
+    columnwise = {"cols": True, "frame": False}[how]
     return wrap_transformer(
-        estimator, cols, allow_reject=allow_reject, columnwise=columnwise
+        estimator,
+        cols=cols,
+        exclude_cols=exclude_cols,
+        allow_reject=allow_reject,
+        columnwise=columnwise,
     )
 
 
@@ -1080,9 +1086,9 @@ def X(value=NULL):
     Parameters
     ----------
     value : object
-        The value passed to ``skrub.var()``, which is used for previews of the
+        The value passed to :func:`skrub.var()`, which is used for previews of the
         learner's outputs, cross-validation etc. as described in the
-        documentation for ``skrub.var()`` and the examples gallery.
+        documentation for :func:`skrub.var()` and the examples gallery.
 
     Returns
     -------
@@ -1142,9 +1148,9 @@ def y(value=NULL):
     Parameters
     ----------
     value : object
-        The value passed to ``skrub.var()``, which is used for previews of the
+        The value passed to :func:`skrub.var()`, which is used for previews of the
         learner's outputs, cross-validation etc. as described in the
-        documentation for ``skrub.var()`` and the examples gallery.
+        documentation for :func:`skrub.var()` and the examples gallery.
 
     Returns
     -------
@@ -1360,6 +1366,8 @@ class Apply(DataOpImpl):
         "estimator",
         "y",
         "cols",
+        "exclude_cols",
+        "no_wrap",
         "how",
         "allow_reject",
         "unsupervised",
@@ -1397,11 +1405,15 @@ class Apply(DataOpImpl):
 
         if "fit" in method_name:
             cols = yield self.cols
+            exclude_cols = yield self.exclude_cols
+            no_wrap = yield self.no_wrap
             how = yield self.how
             allow_reject = yield self.allow_reject
             self.estimator_ = _wrap_estimator(
                 estimator=estimator,
                 cols=cols,
+                exclude_cols=exclude_cols,
+                no_wrap=no_wrap,
                 how=how,
                 allow_reject=allow_reject,
                 X=X,
@@ -1410,10 +1422,33 @@ class Apply(DataOpImpl):
 
         # 2. Call the appropriate estimator method
 
+        if method_name == "fit" and hasattr(self.estimator_, "fit_transform"):
+            # We are a transformer in 'fit' mode. Rather than `fit()` we call
+            # `fit_transform()`. This is done even if the transformer is the
+            # last estimator, as it can make things easier for downstream nodes
+            # if we are building a transformer rather than a predictor.
+            method_name = "fit_transform"
+
+        if "transform" not in method_name:
+            # We are evaluating in another mode than transform or fit_transform
+            # and the current estimator has the corresponding method. We check
+            # if we are the last estimator: if we are not, we want to transform
+            # instead so that subsequent steps get the transformed data rather
+            # than, for example, a fitted estimator or a score.
+            #
+            # An example situation where this arises is are in score mode and
+            # we have an internal transformer such as PCA that has a score()
+            # method.
+            from ._evaluation import HasRunningApplyAncestor
+
+            has_running_apply_ancestor = yield HasRunningApplyAncestor()
+            if has_running_apply_ancestor:
+                method_name = "fit_transform" if "fit" in method_name else "transform"
+
         if "transform" in method_name and not hasattr(self.estimator_, method_name):
-            # We are a predictor and the mode is 'transform' or 'fit_transform' (as in
-            # `.skb.preview()` or `.skb.eval()`). We replace `.transform()`
-            # with `.predict()`
+            # We are a predictor and need to do 'transform' or 'fit_transform'
+            # (as in `.skb.preview()` or `.skb.eval()`). We replace
+            # `.transform()` with `.predict()`
             if method_name == "fit_transform":
                 fit_kwargs = yield from self._eval_kwargs("fit")
                 self.estimator_.fit(X, y, **fit_kwargs)
@@ -1422,11 +1457,6 @@ class Apply(DataOpImpl):
             # In `(fit_)transform` mode only, format the predictions as a
             # dataframe or column if y was one during `fit()`
             return self._format_predictions(X, pred)
-
-        if method_name == "fit" and hasattr(self.estimator_, "fit_transform"):
-            # We are a transformer in 'fit' mode. Rather than `fit()` we call
-            # `fit_transform()` so that subsequent steps can be fitted as well.
-            method_name = "fit_transform"
 
         if "fit" in method_name:
             y_arg = () if self.unsupervised else (y,)
@@ -1516,7 +1546,11 @@ class Apply(DataOpImpl):
 
     def __repr__(self):
         estimator = get_chosen_or_default(self.estimator)
-        if estimator.__class__.__name__ in ["ApplyToEachCol", "ApplyToSubFrame"]:
+        if estimator.__class__.__name__ in [
+            "ApplyToEachCol",
+            "ApplyToSubFrame",
+            "ApplyToCols",
+        ]:
             estimator = estimator.transformer
         # estimator can be None or 'passthrough'
         if isinstance(estimator, str):
@@ -1971,7 +2005,7 @@ def eval_mode():
     This can be:
 
     - 'preview': when the previews are being eagerly computed when the
-      DataOp is defined or when we call ``.skb.eval()`` without
+      DataOp is defined or when we call :func:`DataOp.skb.eval()` without
       arguments.
     - otherwise, the method we called on the learner such as ``'predict'``
       or ``'fit_transform'``.