+++ skrub/_data_ops/tests/test_data_ops.py
@@ -1,4 +1,5 @@
 import copy
+import re
 import sys
 import warnings
 
@@ -8,9 +9,10 @@ import pytest
 from pandas.testing import assert_frame_equal
 from sklearn.base import BaseEstimator
 from sklearn.datasets import make_classification, make_regression
+from sklearn.decomposition import PCA
 from sklearn.dummy import DummyRegressor
 from sklearn.linear_model import LogisticRegression, Ridge
-from sklearn.model_selection import train_test_split
+from sklearn.model_selection import GroupKFold, train_test_split
 from sklearn.utils import check_random_state
 
 import skrub
@@ -197,6 +199,36 @@ def test_predictor_as_transformer():
     expected = pd.DataFrame({"a": [2.0, 2.0, 2.0], "b": [20.0, 20.0, 20.0]})
     assert_frame_equal(learner.fit_transform({"X": X, "y": X}), expected)
     assert_frame_equal(learner.transform({"X": X, "y": X}), expected)
+
+
+def test_transformer_with_score():
+    """
+    When a transformer is the last estimator, we can call score() on it, but
+    when it is the descendant of other estimators it does a (fit_)transform
+    even in other modes such as score.
+
+    What we consider is whether it is the last _estimator_ to be evaluated in
+    the current run, other nodes such as a Choice wrapping the final estimators
+    or other nodes than Apply do not affect this.
+    """
+    X_a, y_a = make_regression()
+    env = {"X": X_a, "y": y_a}
+    feat = skrub.X().skb.apply(PCA(n_components=2)).skb.apply_func(lambda x: x)
+    y = skrub.y()
+    ridge_1 = feat.skb.apply(Ridge(), y=y)
+    ridge_2 = feat.skb.apply(Ridge(), y=y)
+    pred = skrub.choose_from([ridge_1, ridge_2]).as_data_op()
+    cv = pred.skb.cross_validate(env)
+    assert not cv["test_score"].isna().any()
+
+    transformer = feat.skb.make_learner().fit(env)
+    # The PCA is the last estimator: it performs score()
+    assert isinstance(transformer.score(env), float)
+
+    predictor = pred.skb.make_learner().fit(env)
+    # The PCA is not the last estimator: it performs transform()
+    # if it performed score() instead the next (Ridge) step would fail.
+    predictor.score(env)
 
 
 def test_predictor_outputs():
@@ -496,17 +528,17 @@ def test_data_op_impl():
         a.skb.eval()
 
 
-@pytest.mark.parametrize("why_no_wrap", ["numpy", "predictor", "how"])
-@pytest.mark.parametrize("bad_param", ["cols", "how", "allow_reject"])
+@pytest.mark.parametrize("why_no_wrap", ["numpy", "predictor", "no_wrap", "how"])
+@pytest.mark.parametrize("bad_param", ["cols", "exclude_cols", "how", "allow_reject"])
 def test_apply_bad_params(why_no_wrap, bad_param):
     # When the estimator is a predictor or the input is a numpy array (not a
-    # dataframe) (or how='no_wrap') the estimator can only be applied to the
-    # full input without wrapping in ApplyToEachCol or ApplyToSubFrame. In this case
-    # if the user passed a parameter that would require wrapping, such as
-    # passing a value for `cols` that is not `all()`, or passing
-    # how='cols' or allow_reject=True, we get an error.
+    # dataframe) (or no_wrap=True, or how='no_wrap') the estimator can only be
+    # applied to the full input without wrapping in ApplyToCols, ApplyToEachCol
+    # or ApplyToSubFrame. In this case if the user passed a parameter that
+    # would require wrapping, such as passing a value for `cols` that is not
+    # `all()`, or passing how='cols' or allow_reject=True, we get an error.
 
-    if why_no_wrap == bad_param == "how":
+    if bad_param == "how" and why_no_wrap in ["no_wrap", "how"]:
         return
     X_a, y_a = make_classification(random_state=0)
     X_df = pd.DataFrame(X_a, columns=[f"col_{i}" for i in range(X_a.shape[1])])
@@ -528,31 +560,72 @@ def test_apply_bad_params(why_no_wrap, bad_param):
             cols = ["col_0", "col_1"]
     else:
         cols = s.all()
+    if bad_param == "exclude_cols":
+        if why_no_wrap == "numpy":
+            exclude_cols = [0]
+        else:
+            exclude_cols = ["col_0"]
+    else:
+        exclude_cols = None
     how = "cols" if bad_param == "how" else how
     allow_reject = True if bad_param == "allow_reject" else False
+    no_wrap = True if why_no_wrap == "no_wrap" else False
 
     with pytest.raises(
         (ValueError, RuntimeError),
         match=(
-            r"(`cols` must be `all\(\)`|`how` must be 'auto'|`allow_reject` must be"
-            r" False)"
+            r"(`cols` must be `all\(\)`|`exclude_cols` must be None|"
+            r"`how` must be 'auto'|`allow_reject` must be False)"
         ),
     ):
-        X.skb.apply(estimator, y=y, how=how, allow_reject=allow_reject, cols=cols)
+        with warnings.catch_warnings():
+            warnings.filterwarnings("ignore", message=".*how.*deprecated.*")
+            X.skb.apply(
+                estimator,
+                y=y,
+                no_wrap=no_wrap,
+                how=how,
+                allow_reject=allow_reject,
+                cols=cols,
+                exclude_cols=exclude_cols,
+            )
 
 
-def test_apply_invalid_how():
+def test_apply_how():
     df = pd.DataFrame({"a": [1, 2, 3], "b": [4, 5, 6]})
     X = skrub.var("X", df)
     t = PassThrough()
-    for how in ["auto", "cols", "frame", "no_wrap"]:
-        assert list(X.skb.apply(t, how=how).skb.eval().columns) == ["a", "b"]
+    for how in ["cols", "frame", "no_wrap"]:
+        with pytest.warns(
+            FutureWarning,
+            match=re.escape("The 'how' parameter of .skb.apply() has been deprecated"),
+        ):
+            assert list(X.skb.apply(t, how=how).skb.eval().columns) == ["a", "b"]
+    assert list(X.skb.apply(t, how="auto").skb.eval().columns) == ["a", "b"]
     with pytest.raises(RuntimeError, match="`how` must be one of"):
         X.skb.apply(t, how="bad value")
-    # TODO: remove when old names are dropped in 0.7.0
-    with pytest.warns(FutureWarning, match="'columnwise' has been renamed to 'cols'"):
-        wrapper = X.skb.apply(t, how="columnwise").skb.applied_estimator.skb.eval()
+    with pytest.warns(
+        FutureWarning,
+        match=re.escape("The 'how' parameter of .skb.apply() has been deprecated"),
+    ):
+        wrapper = X.skb.apply(t, how="cols").skb.applied_estimator.skb.eval()
         assert isinstance(wrapper, ApplyToEachCol)
+
+
+def test_apply_no_wrap():
+    df = pd.DataFrame({"a": [1, 2, 3], "b": [4, 5, 6]})
+    X = skrub.var("X", df)
+    t = PassThrough()
+    with pytest.raises(
+        RuntimeError,
+        match=re.escape("The parameter 'no_wrap' of .skb.apply() must be a Boolean"),
+    ):
+        X.skb.apply(t, no_wrap="bad")
+    applied = X.skb.apply(t).skb.applied_estimator.skb.eval()
+    assert isinstance(applied, skrub.ApplyToCols)
+    assert isinstance(applied.transformer_, PassThrough)
+    applied = X.skb.apply(t, no_wrap=True).skb.applied_estimator.skb.eval()
+    assert isinstance(applied, PassThrough)
 
 
 class Mul(BaseEstimator):
@@ -635,9 +708,13 @@ def test_apply_transformer_kwargs():
         .skb.make_learner()
     )
     df = pd.DataFrame({"A": [1, 2, 3], "B": [4, 5, 6]})
-    learner.fit({"df": df})
-    learner.fit_transform({"df": df})
-    learner.transform({"df": df})
+    with pytest.warns(
+        FutureWarning,
+        match=re.escape("The 'how' parameter of .skb.apply() has been deprecated"),
+    ):
+        learner.fit({"df": df})
+        learner.fit_transform({"df": df})
+        learner.transform({"df": df})
 
 
 def test_apply_kwargs_evaluation():
@@ -907,3 +984,49 @@ def test_copy_attrs():
         .skb.set_name("transform")
     )
     assert isinstance(out.skb.applied_estimator.skb.eval().transformer_, PassThrough)
+
+
+def test_find():
+    a = skrub.var("a")
+    b = skrub.var("b")
+    c = skrub.choose_from([1, 2], name="c")
+    d = (a + b).skb.set_name("d")
+    e = c + d
+    assert e.skb.find("c") is c
+    assert e.skb.find(lambda n: not hasattr(n, "skb") and n.name == "c") is c
+    assert e.skb.find("d") is d
+    assert e.skb.find(lambda n: hasattr(n, "skb") and n.skb.name == "d") is d
+    assert e.skb.find("z") is None
+    assert e.skb.find(lambda n: False) is None
+    with pytest.raises(TypeError, match="what should either be a string or a callable"):
+        e.skb.find(c)
+    with pytest.raises(TypeError, match="what should either be a string or a callable"):
+        e.skb.find(0)
+
+
+def test_find_X_y():
+    X = skrub.X()
+    y = skrub.y()
+    pred = X.skb.apply(DummyRegressor(), y=y)
+    Xy = pred.skb.find_X_y()
+    assert list(Xy) == ["X", "y"]
+    assert Xy["X"] is X
+    assert Xy["y"] is y
+    Xy = X.skb.find_X_y()
+    assert list(Xy) == ["X"]
+    assert Xy["X"] is X
+    Xy = y.skb.find_X_y()
+    assert list(Xy) == ["y"]
+    assert Xy["y"] is y
+    assert skrub.var("a").skb.find_X_y() == {}
+    groups = skrub.var("groups")
+    kfold = GroupKFold()
+    X = skrub.var("X").skb.mark_as_X(cv=kfold, split_kwargs={"groups": groups})
+    y = skrub.y()
+    pred = X.skb.apply(DummyRegressor(), y=y)
+    Xy = pred.skb.find_X_y()
+    assert list(Xy) == ["X", "cv", "split_kwargs", "y"]
+    assert Xy["X"] is X
+    assert Xy["y"] is y
+    assert Xy["cv"] is kfold
+    assert Xy["split_kwargs"]["groups"] is groups